[fix] 회비 api와 context 분리
