+++ src/hooks/DuesContext.js
@@ -1,7 +1,6 @@
 /* eslint-disable react/jsx-no-constructed-context-values */
 // src/contexts/ApiContext.js
-import React, { createContext, useState, useEffect } from 'react';
-import axios from 'axios';
+import React, { createContext, useState } from 'react';
 
 export const DuesContext = createContext();
 
@@ -11,36 +10,19 @@ export const DuesProvider = ({ children }) => {
   const [totalAmount, setTotalAmount] = useState(0);
   const [currentAmount, setCurrentAmount] = useState(0);
   const [myCardinal, setCardinal] = useState(0);
-  const accessToken = localStorage.getItem('accessToken');
-  useEffect(() => {
-    try {
-      const cardinal = 3;
-      const headers = {
-        Authorization: `Bearer ${accessToken}`,
-      };
-      const BASE_URL = process.env.REACT_APP_BASE_URL;
-      const response = axios.get(`${BASE_URL}/account/${cardinal}`, {
-        headers,
-      });
-      const result = response.data;
-      if (result.code === 200) {
-        if (result.data.cardinal !== myCardinal) {
-          setDuesData(result.data.receipts);
-          setTotalAmount(result.data.total);
-          setCardinal(result.data.cardinal);
-          setCurrentAmount(result.data.currentAmount);
-        }
-      } else {
-        console.error('Failed to get data:', result.message);
-      }
-    } catch (error) {
-      console.error('Error getting data:', error);
-    }
-  }, [accessToken]);
 
   return (
     <DuesContext.Provider
-      value={{ duesData, totalAmount, currentAmount, myCardinal }}
+      value={{
+        duesData,
+        setDuesData,
+        totalAmount,
+        setTotalAmount,
+        currentAmount,
+        setCurrentAmount,
+        myCardinal,
+        setCardinal,
+      }}
     >
       {children}
     </DuesContext.Provider>
