+++ src/utils/handleLogout.js
@@ -15,7 +15,7 @@ const handleLogout = () => {
     if (window.confirm('로그아웃 하시겠습니까?')) {
       localStorage.removeItem('accessToken');
       localStorage.removeItem('refreshToken');
-      navigate('/login');
+      navigate('/');
     }
   };
